feat: auto refresh text editor content onResume() (when app regains focus)
This feature helps a lot with background cloud synchronization services, such as Syncthing or Dropbox. Essentially, the app will auto reload the latest version of the text when not in focus (but if the app is still in focus, the user must manually reload the current file to accept the new changes).

Signed-off-by: Stephen L. <LRQ3000@gmail.com>